Use federation-link-util to extract and implement federation tags dur… (#6484)
