+++ packages/services/schema/package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "@apollo/federation": "0.38.1",
     "@graphql-hive/external-composition": "workspace:*",
+    "@graphql-hive/federation-link-utils": "workspace:*",
     "@graphql-tools/stitch": "9.4.13",
     "@graphql-tools/stitching-directives": "3.1.24",
     "@hive/service-common": "workspace:*",
