+++ packages/services/schema/src/lib/federation-tag-extraction.spec.ts
@@ -2,67 +2,8 @@ import { parse, print } from 'graphql';
 import {
   applyTagFilterOnSubgraphs,
   applyTagFilterToInaccessibleTransformOnSubgraphSchema,
-  extractTagsFromFederation2SupergraphSDL,
-  getFederationTagDirectiveNameForSubgraphSDL,
-  getTagDirectiveNameFromFederation2SupergraphSDL,
   type Federation2SubgraphDocumentNodeByTagsFilter,
 } from './federation-tag-extraction';
-
-describe('getFederationTagDirectiveNameForSubgraphSDL', () => {
-  test('subgraph specification without link directive (Federation 1) -> "tag"', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema {
-        query: Query
-      }
-    `);
-
-    expect(getFederationTagDirectiveNameForSubgraphSDL(sdl)).toEqual('tag');
-  });
-
-  test('subgraph specification with link subgraph directive (Federation 2) -> "federation__tag"', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema @link(url: "https://specs.apollo.dev/federation/v2.0") {
-        query: Query
-      }
-    `);
-
-    expect(getFederationTagDirectiveNameForSubgraphSDL(sdl)).toEqual('federation__tag');
-  });
-
-  test('subgraph specification with link subgraph directive granular import (Federation 2) -> "tag"', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema @link(url: "https://specs.apollo.dev/federation/v2.0", import: ["@tag"]) {
-        query: Query
-      }
-    `);
-
-    expect(getFederationTagDirectiveNameForSubgraphSDL(sdl)).toEqual('tag');
-  });
-
-  test('subgraph specification with link subgraph directive granular import object (Federation 2) -> "tag"', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema @link(url: "https://specs.apollo.dev/federation/v2.0", import: [{ name: "@tag" }]) {
-        query: Query
-      }
-    `);
-
-    expect(getFederationTagDirectiveNameForSubgraphSDL(sdl)).toEqual('tag');
-  });
-
-  test('subgraph specification with link subgraph directive granular import object lias (Federation 2) -> "label"', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(
-          url: "https://specs.apollo.dev/federation/v2.0"
-          import: [{ name: "@tag", as: "@label" }]
-        ) {
-        query: Query
-      }
-    `);
-
-    expect(getFederationTagDirectiveNameForSubgraphSDL(sdl)).toEqual('label');
-  });
-});
 
 describe('applyTagFilterToInaccessibleTransformOnSubgraphSchema', () => {
   describe('correct @inaccessible directive usage based on subgraph version/imports', () => {
@@ -1718,6 +1659,9 @@ describe('applyTagFilterOnSubgraphs', () => {
       exclude: new Set(),
     };
     const typeDefs = parse(/* GraphQL */ `
+      # @note requires federation to be linked now.
+      # extend schema @link(url: "https://specs.apollo.dev/federation/v2.3")
+
       type Query {
         field1: String!
         field2: Type1!
@@ -2185,96 +2129,5 @@ describe('applyTagFilterOnSubgraphs', () => {
         B
       }
     `);
-  });
-});
-
-describe('getTagDirectiveNameFromFederation2SupergraphSDL', () => {
-  test('supergraph specification without import argument -> null', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION) {
-        query: Query
-      }
-    `);
-
-    expect(getTagDirectiveNameFromFederation2SupergraphSDL(sdl)).toEqual(null);
-  });
-
-  test('supergraph import without alias -> tag', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
-        @link(url: "https://specs.apollo.dev/tag/v0.3") {
-        query: Query
-      }
-    `);
-
-    expect(getTagDirectiveNameFromFederation2SupergraphSDL(sdl)).toEqual('tag');
-  });
-
-  test('supergraph import with alias -> alias name', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
-        @link(url: "https://specs.apollo.dev/tag/v0.3", as: "federation__tag") {
-        query: Query
-      }
-    `);
-
-    expect(getTagDirectiveNameFromFederation2SupergraphSDL(sdl)).toEqual('federation__tag');
-  });
-});
-
-describe('extractTagsFromFederation2SupergraphSDL', () => {
-  test('supergraph specification without import argument -> null', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION) {
-        query: Query
-      }
-      type Query {
-        foo: String!
-      }
-    `);
-
-    expect(extractTagsFromFederation2SupergraphSDL(sdl)).toEqual(null);
-  });
-
-  test('supergraph import without alias -> tag', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
-        @link(url: "https://specs.apollo.dev/tag/v0.3") {
-        query: Query
-      }
-      type Query {
-        foo: String! @tag(name: "foo")
-        kekeke: String! @tag(name: "kekeke")
-      }
-    `);
-
-    expect(extractTagsFromFederation2SupergraphSDL(sdl)).toEqual(['foo', 'kekeke']);
-  });
-
-  test('supergraph import with alias -> federation__tag', () => {
-    const sdl = parse(/* GraphQL */ `
-      schema
-        @link(url: "https://specs.apollo.dev/link/v1.0")
-        @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
-        @link(url: "https://specs.apollo.dev/tag/v0.3", as: "federation__tag") {
-        query: Query
-      }
-      type Query {
-        foo: String! @federation__tag(name: "foo")
-        kekeke: String!
-      }
-    `);
-
-    expect(extractTagsFromFederation2SupergraphSDL(sdl)).toEqual(['foo']);
   });
 });
